+++ modules/ui/sections/validation_issues.js
@@ -10,6 +10,7 @@ import { prefs } from '../../core/preferences';
 import { t } from '../../core/localizer';
 import { utilHighlightEntities } from '../../util';
 import { uiSection } from '../section';
+import { validationIssue } from '../../core/validation';
 
 export function uiSectionValidationIssues(id, severity, context) {
 
@@ -105,7 +106,7 @@ export function uiSectionValidationIssues(id, severity, context) {
             .append('span')
             .attr('class', 'issue-icon')
             .each(function(d) {
-                var iconName = '#iD-icon-' + (d.severity === 'warning' ? 'alert' : 'error');
+                const iconName = `#iD-icon-${validationIssue.ICONS[d.severity]}`;
                 d3_select(this)
                     .call(svgIcon(iconName));
             });